Optimize DISCOVERING_HOSTS_WITH_TCPSYNCHRONIZATION function
This commit modifies the DISCOVERING_HOSTS_WITH_TCPSYNCHRONIZATION function to use the -oG option of nmap and xargs to pass the list of hosts as arguments, which should improve the performance of the function. The grep and cut commands have also been modified to only run once, and the output of the nmap command is now logged to a file and displayed on the screen using tee. These changes should make the function faster and more efficient.